Se crean los trigger y se inicia la conexión con Java

+++ Sentencias SQL/Creación tablas/Creación animal, dieta y animal_dieta.sql
@@ -7,7 +7,7 @@ create table dieta(
 create table animal(
 	idAnimal int auto_increment primary key,
     nombre varchar(60),
-    peso int,
+    peso float,
     clasificacion varchar(50)
 );
